+++ SteamPS/Public/Install-SteamCMD.ps1
@@ -23,6 +23,9 @@ function Install-SteamCMD {
     .NOTES
     Author: Frederik Hjorslev Poulsen
 
+    .LINK
+    https://hjorslev.github.io/SteamPS/Install-SteamCMD.html
+
     #>
 
     [CmdletBinding()]