+++ build.ps1
@@ -13,7 +13,7 @@ if ($env:APPVEYOR_REPO_BRANCH -ne 'master') {
 } else {
     # We're going to add 1 to the revision value since a new commit has been merged to Master
     # This means that the major / minor / build values will be consistent across GitHub and the Gallery
-    Try {
+    try {
         # This is where the module manifest lives
         $manifestPath = ".\$SteamPSModule\$SteamPSModule.psd1"
 
@@ -46,10 +46,12 @@ if ($env:APPVEYOR_REPO_BRANCH -ne 'master') {
     Import-Module -Name "$PSScriptRoot\$SteamPSModule" -Force
     New-MarkdownHelp -Module $SteamPSModule -OutputFolder '.\docs\' -Force
     New-ExternalHelp -Path '.\docs\' -OutputPath ".\docs\en-US\" -Force
+    Copy-Item -Path '.\README.md' -Destination 'docs\index.md'
+    Copy-Item -Path '.\CHANGELOG.md' -Destination 'docs\CHANGELOG.md'
     Write-Host -Object ''
 
     # Publish the new version to the PowerShell Gallery
-    Try {
+    try {
         # Build a splat containing the required details and make sure to Stop for errors which will trigger the catch
         $PM = @{
             Path        = ".\$SteamPSModule"
@@ -59,14 +61,14 @@ if ($env:APPVEYOR_REPO_BRANCH -ne 'master') {
 
         Publish-Module @PM
         Write-Host "$SteamPSModule PowerShell Module version $newVersion published to the PowerShell Gallery." -ForegroundColor Cyan
-    } Catch {
+    } catch {
         # Sad panda; it broke
         Write-Warning "Publishing update $newVersion to the PowerShell Gallery failed."
         throw $_
     }
 
     # Publish the new version back to Master on GitHub
-    Try {
+    try {
         # Set up a path to the git.exe cmd, import posh-git to give us control over git, and then push changes to GitHub
         # Note that "update version" is included in the appveyor.yml file's "skip a build" regex to avoid a loop
         $env:Path += ";$env:ProgramFiles\Git\cmd"
@@ -77,7 +79,7 @@ if ($env:APPVEYOR_REPO_BRANCH -ne 'master') {
         git commit -s -m "Update version to $newVersion"
         git push origin master
         Write-Host "$SteamPSModule PowerShell Module version $newVersion published to GitHub." -ForegroundColor Cyan
-    } Catch {
+    } catch {
         # Sad panda; it broke
         Write-Warning "Publishing update $newVersion to GitHub failed."
         throw $_